requirements.txt
selenium~=4.23.1
SQLAlchemy~=2.0.31
httpx~=0.27.0
uvicorn~=0.30.5
fastapi~=0.112.0
pydantic~=2.8.2
python-dotenv